weblate/utils/scripts.py
# -*- coding: utf-8 -*-
#
# Copyright © 2012 - 2017 Michal Čihař <michal@cihar.com>
#
# This file is part of Weblate <https://weblate.org/>
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <https://www.gnu.org/licenses/>.
#

import os


def get_script_name(name):
    """Return script name from string possibly containing full path and
    parameters.
    """
    return os.path.basename(name).split()[0]


def get_script_choices(choices):
    """Generate list of script choice in the Admin interface."""
    return [
        (script, get_script_name(script)) for script in choices
    ] + [('', '')]
